Bug 43245 - JSS - Custom Elements - dataset based element could be created as custom ones under Frame Changed the tool name to Composite Element
git-svn-id: http://code.jaspersoft.com/svn/repos/jaspersoftstudio/trunk@5912 6833c83e-4fd0-4c68-a3df-283c6148db30

+++ com.jaspersoft.studio/src/com/jaspersoft/studio/editor/tools/ToolEditWizardPage.java
@@ -45,19 +45,21 @@ public class ToolEditWizardPage extends ToolDefinitionWizardPage {
 	public ToolEditWizardPage(String initialName, String initialDescription, String initialImagePath){
 		this.initialDescription = initialDescription;
 		this.initialName = initialName;
-		
-		File image = new File(initialImagePath);
-		//If exist export the image to the temp directory
-		if (image.exists()){
-			File tmpDir = new File(System.getProperty("java.io.tmpdir")); 
-			File destFile = new File(tmpDir, image.getName());
-			try{
-				FileUtils.copyFile(image, destFile);
-			} catch (Exception ex){
-				ex.printStackTrace();
+
+		if (initialImagePath != null){
+			File image = new File(initialImagePath);
+			//If exist export the image to the temp directory
+			if (image.exists()){
+				File tmpDir = new File(System.getProperty("java.io.tmpdir")); 
+				File destFile = new File(tmpDir, image.getName());
+				try{
+					FileUtils.copyFile(image, destFile);
+				} catch (Exception ex){
+					ex.printStackTrace();
+				}
+				destFile.deleteOnExit();
+				this.initialImagePath = destFile.getAbsolutePath();
 			}
-			destFile.deleteOnExit();
-			this.initialImagePath = destFile.getAbsolutePath();
 		}
 	}
